requirements.txt
fastapi[all]>=0.105.0
uvicorn>=0.24.0
sqlalchemy[mssql]>=2.0.23
python-jose>=3.3.0
pyodbc>=5.0.1
pyjwt>=1.7.1
passlib
bcrypt
